Add new apps for API and Bears
Using `python manage.py startapp [name]`

In Django, each "thing" tends to have an app, or python package that's
more-or-less self-contained.

There aren't really strict rules around this. For example, in this case "API" is an
app that will mostly serve content about bears, so it could just be some
code in the bears app. However, it's possible that we want to add a load
of API-like logic (authentication, user tokens, etc etc) to the API in
future, so I decided to create an app keeping that code tidy.

I've really done this as a way to force the point that apps are
sometimes hard to define. It's more of a social / rough judgement thing
than a set of rules. Most of the time, in real projects, apps will
import between each other. They are coupled, but at least the code can
be sorted in some sort of nice way...!