Move files out of "project" package and in to apps
In Django, there tends to be a "project" package that contains things
like settings, root URLs, templates, project-wide assets, etc.

Django suggests having this package in the same directory as
`manage.py`, and I wonder if this is where the hyphens issue came from?
`manage.py` expects to import this project when Django is run, so it
needs to be a valid python package name. The parent folder to
`manage.py` can be whatever-you-like.

Anyway, in this case we're using `fat_bear_week_api` as the
_project_ name, and that means moving app-specific stuff out in to the
app directories (more on this later).

Note that in DC projects, I like to have all the apps in an `apps`
package inside the project package. (see e.g
https://github.com/DemocracyClub/yournextrepresentative/tree/master/ynr).

This is a Sym-standard, not a Django one. To me, it's nicer, so I've
imposed that on all of DC, but it's not something you have to do, so
I'll not do it for this project in order to keep it more like a stock
Django project.