Remove Django managed models, add migrations
Let's get in to the ORM :tada:

So, Django manages the database for you. This is in a lot of ways great,
but it might involve thinking backwards if you're used to modelling in
the database and then writing code to work with it.

The models here are (more or less) how you would _start_ in Django,
before you have any data. That is, you would define the models and
relationships in code before you touch the database.

You then run `python manage.py makemigrations` in order to get Django to
write the code that can be converted to SQL and be run on your database.
`makemigrations` doesn't do anything with the database!

After you have some migrations, you can run `migrate` to do the work of
changing the database.

Migrations are mainly about schema rather than data, but you can add or
change data in a data migration.

For example, if you always wanted to have a user group in the database,
you could add a data migration to populate the database with it each
time a project is installed. We don't need them right now, so let's
assume we're just talking about schema changes.