Remove a bunch of stuff from local settings template
We only want stuff in here that we want to change on per-install. Things
like `INSTALLED_APPS` shouldn't live here, as they are project settings.

Of course there are many ways to do local config. Some people like
taking from `os.envion` for everything (with defaults), and managing
that using a `.env` file. This is a good system, but for now let's just
make this file smaller