Replace BearView with BearViewSet
A few things going on here.

1. I've moved the URLs around a bit. We're now using a DRF "Router".
This is some magic that knows how to route URLs for us. It will take
care of things like allowing POST for some users and setting up
list/detail URLs for our viewsets. There's a lot in Django/Python that
isn't magic, but this is one area where it's quite useful.

2. I've changed `BearView` to `BearViewSet`. This is a class that knows
about common view patterns, mainly, list/detail views. We can do more
with them, but out of the box we can give it a QuerySet and a
serializer_class, and it will do the rest for us.

3. I've moved the router in to the API app, and used an include to let
that app manage it's own URLs. The app is _mounted_ at `/api/` but from
this point on the core project doesn't need to know about how the app
wants to deal with sub-URLs. This is one nice thing about the app
concept.