Use a manger for complex queries
It's a good idea to try to keep DB logic tied to the models, and view
logic in the view later. This means, every time you see a complex
database query in a view, you can consider refactoring it in to the
model later.

In this case I've created a new "manager". This allows us to extend the
default manager "Model.objects.all()" to add new methods.

Here I didn't bother thinking about what the logic should be, I just
added a placeholder and a TODO ;)

Note that I didn't convert these to a DRF ViewSet as we only need a
single view, not a list/detail view